requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
requests==2.31.0
httpx==0.25.0
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
